requirements.txt
coverage
python-coveralls
pytest-cov
codecov
numpy
pandas
scikit-learn
matplotlib
seaborn
statsmodels
astropy
